Math::Trig: correct interval in documentation
This effectively undoes commit 02b54f9d07, which tried to fix a typo,
but introduced a technical inaccuracy because it wasn't a typo.

The intent is to specify a half-open interval including 0 but not
including 2pi. The original notation used was [a, b[ which is common in
some European countries, but [a, b) seems to be more common in the
English-speaking world.