utf8.c: Move most important conditional to be first
It turns out that the information generated in this block is only needed
if the final conditional in this complicated group of them is true,
which checks if the caller wants anything special for certain classes of
code points.  Because that final condition is subsidiary, the block was
getting executed just to be thrown away.