utf8.c: Check specially for perl-extended UTF-8
More rigorous testing of the overlong malformation, yet to be committed,
showed that this needs to be handled specially.  This commit does part
of that.

Perl extended UTF-8 means you are using a start byte not recognized by
any UTF-8 standard.  Suppose it is an overlong sequence that reduces
down to something representable using standard UTF-8.  The string still
used non-standard UTF-8 to get there, so should still be called out when
the input parameters to this function ask for that.  This commit is a
first step towards that.