utf8.c: Combine two blocks
It turns out that the work being done in the first block is only used in
the second block.  If that block doesn't get executed, the first block's
effort is thrown away.  So fold the first block into the second.  This
results in a bunch of temporaries that were used to communicate between
the blocks being able to be removed.

More detailed comments are added.