utf8_to_uv_msgs: Hoist common paradigm out of switch
Previously this variable was initialized to 0 and then set to another
value in each case in the switch.  But most of the cases set it to the
same value.  Might as well initialize it to that value, remove the
statements that merely repeat that, and leave in the cases that set it
to something else.