podcheck.t: Any module we find is valid to link to
podcheck looks through the source distribution for pod.  It skips
internal checking of ones in /cpan (unless a command line option says
to.  If you do this, be prepared for finding gobs of problems in
these!).

But any such module found is fair game to be linked to by modules we do
examine.  It is a simple matter to add them to the list of known modules
that can be linked to.

Prior to this commit this was done surreptitiously by extending the
Pod::Checker node() method.  But a future commit will take this away.