utf8n_to_uv_msgs: Avoid unnecessary work
The outermost block here is executed when any of three types of
problematic Unicode code points is encountered, and the caller has
indicated special handling of at least one of those types.  Before this
commit, we set a flag to later look to see if what was encountered
matched the type the caller specified.  This commit changes to do that
looking at the point where the flag had been set, and only sets the flag
if necessary.  This may completely avoid the later work, which has
set-up overhead, and this will make future commits simpler.