utf8_to_uv_msgs: Reorder case statements in switch()
This just moves these around so that they are roughly in increasing
order of complexity, which makes it easier to grok.

This doesn't change the order the cases are called in when there are
multiple ones.  The order is dictated by what the switch() does first,
which isn't affected by this commit.