utf8.c: Create a #define
This creates a #define of the string in a string array constant.  The
constant was to avoid having this common string appear in multiple
places, but there is a benefit to having it in one place separately,
which this commit also does.  And that is, it makes a call to
Perl_form() consistent with two other calls that the next commit will
combine.