toke.c: minor adjustment of #ifdef to stop confusing code editors
Before this change, the two separate `while` statements would upset the
code folding as parsed by tree-sitter-c (and likely many others, I
haven't tested), into thinking this was two nested loops. Having failed
to find the end of both of them before the end of the function, various
confusions result, usually ending up in the entire rest of the file (and
it's a long file) getting folded into one giant region. This likely
causes various static analysis tools similarly to not see any of the
subsequent functions in the file.

By adjusting the code so that just the condition part is conditional on
the `#ifdef`, it means that code parsing tools have a much easier time
working out the high-level structure of this file.