utf8_to_uv_msgs: Move code to ensure initialization
There was a path through this function in which the caller's parameter
it asked to be set, &msgs, did not get set.  And doing it at the
beginning means not needing a second place.

Similarly for &errors.  There is no path where it didn't get set, but it
is cleaner to do it in at the same time as doing msgs.