Configure: Fix compilation error on some systems
See https://github.com/Perl/perl5/issues/22793

This typo in a printf format causes some C compilations to fail; others
to just warn.  However stderr is redirected to /dev/null, so we weren't
aware of this issue.

When it fails, other issues cause perl to not be usable on the system.