Add two new flags to utf8_to_uv_msgs()
These new flags subsume the functionality of
force_out_malformed_utf8_message().  The guts of that function are moved
to this one.

This commit resulted from my observation that a bunch of code that
called some form of utf8_to_uv() would test the success of that, and if
it failed immediately call force_out_malformed_utf8_message().  It's
better to do a common paradigm in just one place, and this commit paves
the way for that.