utf8.h Swap which definition is the base
The GOT defines are now the principal ones, and the ALLOW ones are
changed to be in terms of the GOT.  This seems right to me, as the GOT
ones are the ones anyone might actually care about.  (The ALLOW ones are
for weird cases.)  More importantly this makes the next commit make more
sense.