openbsd: ensure we link to the built libperl.a, not the system libperl.a
When building with gcc, lib/ExtUtils/t/Embed.t would link against
the system libperl.a rather than the newly built libperl.a.

Due to the limited API used by the sample code this typically didn't
crash, but some configuration changes could result in a crash or a
link error.

For OpenBSD, change the link options to more closely match those
used when building the perl executable, which results in linking
against the correct library.

Fixes #22125