op.c/peep.c - Optimise away empty else{} blocks
This is a replacement for https://github.com/Perl/perl5/pull/22745, which
aimed to prevent the associated `OP_STUB` from being created in
`Perl_newCONDOP` and subsequently removed in peep.c. However, doing that
turned out to be unavoidable, since it is not safe to omit the `OP_STUB`
if it is in scalar context, but `Perl_newCONDOP` runs before context is
applied.

This commit:
 * Re-adds the logic in `Perl_op_scope` to not wrap a bare `OP_STUB` in an
   `ENTER/LEAVE` pair.
 * Removes any relevant `OP_STUB`s in non-scalar context in `Perl_rpeep`.
 * Adds B::Deparse support
 * Adds a test based on the failures seen in https://github.com/Perl/perl5/issues/22866