op.h/op.c/peep.c - Optimise away empty if{} blocks
This commit optimises away the OP_STUB associated with an empty `true`
branch of an `OP_COND_EXPR`.

The `OPf_SPECIAL` flag has been brought into use on the `OP_COND_EXPR`
to indicate that it is the `else`, not the `if`, block that has been
optimised away. This is purely for the benefit of B::Deparse.