Simplify the recommended alternatives to rand()
The CPAN Security Group (CPANSec) is currently working on guides to
generating security-quality random data [1]. We are focusing on modules
that have secure defaults and are fairly lightweight.

We would like to change the recommended modules to ones that we think
are better options.

Crypt::URandom is portable, has fewer prerequisites than Crypt::Random,
Math::Random::Secure or Data::Entropy, and works with Windows.

Crypt::PRNG has secure defaults and methods for generating different
kinds of random data.

Math::TrulyRandom hasn't been updated since 1996, and is a solution
intended for systems without something like /dev/random.

Data::Entropy had a security issue that was recently fixed, and the
latest version is marked as deprecated.

[1] https://security.metacpan.org/docs/guides/random-data-for-security.html