Data-Dumper: fix deparse.t for non-XS testing
This is a follow-up to b5048e7b9abae986778038f1e36281c6175b17e1, which
added 'Deparse' support to XS and made both XS and pure Perl testing
mandatory. The other test files treat XS as optional, so this patch
follows that pattern. (If we decide that the XS part of Data::Dumper is
not optional, we should change all the other test files for
consistency.)

Previously, testing without XS available resulted in:

    t/deparse.t ................... Undefined subroutine &Data::Dumper::Dumpxs called at .../dist/Data-Dumper/blib/lib/Data/Dumper.pm line 213.
    # Looks like your test exited with 255 before it could output anything.
    t/deparse.t ................... Dubious, test returned 255 (wstat 65280, 0xff00)
    Failed 16/16 subtests
    ...
    t/deparse.t                 (Wstat: 65280 (exited 255) Tests: 0 Failed: 0)
      Non-zero exit status: 255
      Parse errors: Bad plan.  You planned 16 tests but ran 0.