Data-Dumper: fix bugs.t for non-XS testing
This is a follow-up to 7d3a730ee869d89a6f40963c80aaa0e044b0c7d2, which
fixed an out-of-bounds write in Dumpxs when passed invalid UTF-8. This
test does not apply to the pure Perl implementation (which detects the
invalid input and cleanly throws an error instead).

Previously, testing without XS available would result in:

    t/bugs.t ...................... 1/24 Malformed UTF-8 character (fatal) at .../dist/Data-Dumper/blib/lib/Data/Dumper.pm line 587.
    # Looks like your test exited with 2 just after 6.
    t/bugs.t ...................... Dubious, test returned 2 (wstat 512, 0x200)
    Failed 18/24 subtests
    ...
    t/bugs.t                    (Wstat: 512 (exited 2) Tests: 6 Failed: 0)
      Non-zero exit status: 2
      Parse errors: Bad plan.  You planned 24 tests but ran 6.