Reflect change in CPAN releasor
On CPAN Time-Piece has a new release from a new releasor.  For
consistent results out of 'Porting/core-cpan-diff', we should have
changed the CPANID of the releasor in Maintainers.pl before synching.
We'll correct that here and standardize all $VERSIONs in cpan/Time-Piece
to indicate we're now a bit ahead of CPAN.