mem_collxfrm_: Reorder a few statements
This function has a successful return branch and an unsuccessful return
branch, and a macro common to both, named CLEANUP_STRXFRM.  There is
another code clause repeated in both branches.  This reorders some of the
statements in each branch so that that clause is positioned to just
after the calls to CLEANUP_STRXFRM.  The next commit will add it to the
macro, so as to remove the redundant code clause.