memcollxfrm: Fix a memory leak
This commit creates a new variable, sans_highs, to store newly allocated
memory that only happens sometimes.  Just before returning, that is
freed.

This is the final step in solving the leak spotted by Tony Cook in
https://github.com/Perl/perl5/pull/22811#discussion_r1868542654

The problem was that there were potentially 0 to 3 mallocs in this
function, and it was only freeing up to two of them.  The solution is to
have a separate variable for each malloc, and to free them all before
returning.  If the corresponding malloc did not happen, the variable
will be NULL, and no free will occur.

This makes a loop rather more complicated.  The next commit will
simplify it.