perl.h - rename PERL_DEB() to a more intuitive name
This patch renames PERL_DEB() to PERL_IF_DEBUGGING() and renames
PERL_DEB2() to PERL_IF_DEBUGGING_ELSE(). It also documents them, along
with PERL_DEBUG() and DEBUG_r() and friends.

I have switched the existing PERL_DEB() or PERL_DEB2() calls to use the
new names, but kept aliases for the old names just in case they are used
out in the wild (eg CPAN).

Thanks to mauke for the name suggestions!