Preserve behaviour of allowing an undefined key for tied hash lookups
The previous change prevented the warning, but forced the hash key for
defelem to the empty string if the original key was undefined.

This wasn't a problem if the hash was already tied when the function
was called (and the PVLV created), since the resulting PVLV used
packelem magic instead of defelem.

But if the hash was tied after the PVLV was created references through
the PVLV would be working on $hash{""} instead of $hash{+undef},
breaking this possibly used feature.

One issue is whether we want to allow undef keys for tied hash, if we
do we should probably make them warn less.

Fixes #22423