lib/Unicode/UCD.pm: update num() usage example
A combination of two Rumi digits is not considered a valid decimal
number by num() thus resulting in an undef result value. While such
behaviour is properly documented in the num() description and thus
expected, it might be better to give an example that returns a proper
numeric value.

An equivalent change was made to an example in pod/perlunicook.pod.